+++ src/types/location.ts
@@ -1,0 +1,9 @@
+export type City = Point & {
+    zoom: number;
+};
+
+export type Point = {
+    title: string;
+    latitude: number;
+    longitude: number;
+};
