+++ src/types/offer.ts
@@ -1,0 +1,34 @@
+type OfferLocation = {
+    latitude: number;
+    longitude: number;
+    zoom: number;
+}
+
+export type OfferProps = {
+    id: string;
+    title: string;
+    type: string;
+    price: number;
+    city: {
+        name: string;
+        location: OfferLocation;
+    };
+    location: OfferLocation;
+    isFavorite: boolean;
+    isPremium: boolean;
+    rating: number;
+    previewImage: string;
+}
+
+export type OfferWithDetailsProps = OfferProps & {
+    description: string;
+    bedrooms: number;
+    goods: string[];
+    host: {
+        name: string;
+        avatarUrl: string;
+        isPro: boolean;
+    };
+    images: string[];
+    maxAdults: number;
+};
